+++ locals.tf
@@ -1,0 +1,4 @@
+locals {
+  backup_resource_group_name = var.backup_policy_id != null ? split("/", var.backup_policy_id)[4] : null
+  backup_recovery_vault_name = var.backup_policy_id != null ? split("/", var.backup_policy_id)[8] : null
+}